Added tests for get_metadata func
- Added fixture to return path to create file by prepare_environment and create_file fixtures
