+++ Tests/file_service/test_create_file.py
@@ -1,21 +1,6 @@
 import os
 
-import pytest
-
-from FileServerApp import file_service
 from FileServerApp.config import FILENAME_LEN, FILE_EXTENSION, DEFAULT_FILE_CONTENT
-
-
-@pytest.fixture(scope="module")
-def create_file(prepare_test_environment):
-    # Act
-    filename = file_service.create_file()
-
-    # Assert
-    file_path = os.path.join(prepare_test_environment, filename)
-    assert os.path.isfile(file_path)
-
-    return filename
 
 
 def test_file_creation(prepare_test_environment, create_file):
@@ -29,12 +14,9 @@ def test_filename_after_creation(prepare_test_environment, create_file):
     assert len(create_file) == FILENAME_LEN + len(FILE_EXTENSION)
 
 
-def test_filename_content_after_creation(prepare_test_environment, create_file):
-    # Arrange
-    path = os.path.join(prepare_test_environment, create_file)
-
+def test_filename_content_after_creation(path_to_new_file):
     # Act
-    with open(path, "rt") as r_file:
+    with open(path_to_new_file, "rt") as r_file:
         file_content = r_file.read()
 
     # Assert
